+++ src/client/js/index.js
@@ -16,6 +16,7 @@ import 'datatables.net-buttons/js/buttons.print.mjs';
 import 'datatables.net-colreorder-bs5';
 import 'datatables.net-fixedheader-bs5';
 import 'datatables.net-responsive-bs5';
+import 'datatables.net-rowgroup-bs5';
 import 'datatables.net-scroller-bs5';
 //
 import 'datatables.net-bs5/css/dataTables.bootstrap5.min.css';
@@ -23,6 +24,7 @@ import 'datatables.net-buttons-bs5/css/buttons.bootstrap5.min.css';
 import 'datatables.net-colreorder-bs5/css/colReorder.bootstrap5.min.css';
 import 'datatables.net-fixedheader-bs5/css/fixedHeader.bootstrap5.min.css';
 import 'datatables.net-responsive-bs5/css/responsive.bootstrap5.min.css';
+import 'datatables.net-rowgroup-bs5/css/rowGroup.bootstrap5.min.css';
 import 'datatables.net-scroller-bs5/css/scroller.bootstrap5.min.css';
 //
 import '../stylesheets/tabular.less';